requirements.txt
confluent-kafka
yfinance
python-dotenv
pandas
redis
fastapi
uvicorn